feat: add details in monitor
