+++ source/portal/src/pages/chatbot/components/Message.css
@@ -14,9 +14,11 @@
     background-color: #f2f2f2;
     padding: 8px;
     border: 1px solid #ddd;
+    min-width: 120px;
 }
 
 .custom-table-cell {
     padding: 8px;
     border: 1px solid #ddd;
+    min-width: 120px;
 }
